strabo/requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==2.1
psycopg2==2.6.1
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.9.0
Werkzeug==0.10.4
itsdangerous==0.24
requests==2.7.0
SQLAlchemy==1.0.12

## Python libraries ##
geojson==1.3.1
